+++ js/main.js
@@ -1,20 +1,3 @@
 import './form.js';
+import './settings.js'
 import './map.js';
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
